Backport YARN-8108 Patchset-1 to solve "request is replay" issue on RM /proxy URL (#145)
After using cname for resourcemanagers, we observed that the /proxy URL returns a "request is replay" kerberos error. That is due to RMAuthenticationFilter being used twice by specific calls to this URL.
We don't know why this appeared with the cname change, however we found YARN-8108 explaining the double authentication filter issue. The first patch set proposed solves the issue. Hadoop community decided to use patch set 2 to be merged on HDP-3 branch, assuming that the way patchset 2 fixes the issue is better for future code, but does not solve HDP-2 issue without the additional of other properties.

We have tested patchset 1 and it solved the issue without breaking anything else.

Co-authored-by: William Montaz <w.montaz@criteo.com>